# openapi.yaml
openapi: "3.0.3"
info:
  version: 1.0.0
  title: Work Manager
  description: >-
    Customizable work item manager that leverages Camunda workflows (BPMN and DMN), Form.io and customizable data types
    to model online transaction processing.
paths:
  /admin/entity/schema:
    post:
      summary: Create or Update a schema definition.
      operationId: postSchema
      tags: [ 'Schema' ]
      requestBody:
        description: Schema definition.
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SchemaModel'
      responses:
        204:
          description: The Schema was created.
    get:
      summary: Get list of defined schemas with name matching a search string. Omitting search parameter will return all schemas.
      operationId: getSchemas
      tags: [ 'Schema' ]
      parameters:
        - in: query
          name: search
          description: Full or partial schema name to filter returned schemas.
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Collection of schemas matching search criteria.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SchemaModel'
  /admin/entity/schema/{name}:
    get:
      summary: Get a single Schema whose name matches the path exactly.
      operationId: getSchema
      tags: [ 'Schema' ]
      parameters:
        - in: path
          name: name
          description: Name of schema to retrieve
          required: true
          schema:
            type: string
      responses:
        '200':
          description: The requested Schema.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SchemaModel'
        '404':
          description: A schema with the given name does not exist.
  /admin/formconfig:
    post:
      summary: Create or Update a form config definition.
      operationId: postFormConfig
      tags: [ 'FormConfig' ]
      requestBody:
        description: Form Config definition.
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FormConfigDefinitionModel'
      responses:
        200:
          description: The posted form config post-save. This will include assigned IDs for newly created form configs.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FormConfigDefinitionModel'
    get:
      summary: Get list of defined form configs with name matching a search string. Omitting search parameter will return all form configs.
      operationId: getFormConfigs
      tags: [ 'FormConfig' ]
      parameters:
        - in: query
          name: search
          description: Full or partial form config name to filter returned form config.
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Collection of form configs matching search criteria.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/FormConfigDefinitionModel'
  /admin/formconfigID/{id}:
    get:
      summary: Get a single form config whose name matches the path exactly.
      operationId: getFormConfigById
      tags: [ 'FormConfig' ]
      parameters:
        - in: path
          name: id
          description: ID of form config to retrieve
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: The requested form config.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FormConfigDefinitionModel'
        '404':
          description: A form config with the given id does not exist.
  /admin/formconfig/{name}:
    get:
      summary: Get a single form config whose name matches the path exactly.
      operationId: getFormConfigByName
      tags: [ 'FormConfig' ]
      parameters:
        - in: path
          name: name
          description: Name of form config to retrieve
          required: true
          schema:
            type: string
      responses:
        '200':
          description: The requested form config.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FormConfigDefinitionModel'
        '404':
          description: A form config with the given name does not exist.
  /admin/formconfig/publish:
    put:
      summary: Set the status of a given form configuration to published.
      operationId: putPublishFormConfig
      tags: [ 'FormConfig' ]
      requestBody:
        description: The ID of the form configuration to publish.
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FormConfigQueryParameters"
      responses:
        '200':
          description: The form was published successfully.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FormConfigDefinitionModel"
        '404':
          description: Form with the given ID does not exist.
  /admin/formconfig/unpublish:
    put:
      summary: Set the status of a given form configuration to draft.
      operationId: putUnpublishFormConfig
      tags: [ 'FormConfig' ]
      requestBody:
        description: The ID of the form configuration to unpublish.
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FormConfigQueryParameters"
      responses:
        '200':
          description: The form was unpublished successfully.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FormConfigDefinitionModel"
        '404':
          description: Form with the given ID does not exist.
  /admin/transaction:
    post:
      summary: Create or Update a transaction definition.
      operationId: postTransactionDefinition
      tags: [ 'Transaction' ]
      requestBody:
        description: Transaction Definition.
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TransactionDefinitionModel'
      responses:
        200:
          description: The posted transaction definition post-save. This will include assigned IDs for newly created transaction definitions.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionDefinitionModel'
    get:
      summary: Get list of transaction definitions with name matching a search string. Omitting search parameter will return all transaction definitions.
      operationId: getTransactionDefinitions
      tags: [ 'Transaction' ]
      parameters:
        - in: query
          name: name
          description: Full or partial name to filter returned transaction definitions.
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Collection of transaction definitions matching search criteria.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TransactionDefinitionModel'
  /admin/transaction/category:
    get:
      summary: Get list of transaction definitions with name matching a search string
      operationId: getTransactionDefinitionsByCategory
      tags: [ 'Transaction' ]
      parameters:
        - in: query
          name: category
          description: Full or partial name to filter returned transaction definitions by category.
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Collection of transaction definitions matching category criteria.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TransactionDefinitionModel'
  /admin/transaction/{id}:
    get:
      summary: Get a single transaction definition by ID.
      operationId: getTransactionDefinition
      tags: [ 'Transaction' ]
      parameters:
        - in: path
          name: id
          description: ID of transaction definition to retrieve
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: The requested transaction definition.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionDefinitionModel'
        '404':
          description: A schema with the given name does not exist.
  /admin/transaction/linktype:
    post:
      summary: Create or update a transaction link type which can be assigned to a transaction definition.
      operationId: postTransactionLinkType
      tags: [ 'Transaction' ]
      requestBody:
        description: The transaction link type to create.
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TransactionLinkTypeModel'
      responses:
        201:
          description: The created transaction link type.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionLinkTypeModel'
...
  /admin/transaction/allowedlink:
    post:
      summary: Create an allowed link rule for a transaction definition.
      operationId: postAllowedLinkToDefinition
      tags: [ 'Transaction' ]
      requestBody:
        description: The link type to allow on a transaction definition.
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AllowedLinkCreationRequest'
      responses:
        201:
          description: The created transaction link type.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AllowedLinkModel'
  /admin/transaction/allowedlinks/{key}:
    get:
      summary: Get list of allowed links for a transaction definition by transaction definition key.
      operationId: getTransactionDefinitionAllowedLinksByKey
      tags: [ 'Transaction' ]
      parameters:
        - in: path
          name: key
          description: The transaction definition key to search on.
          required: true
          schema:
            type: string
      responses:
        200:
          description: Collection of allowed links for a transaction definition.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AllowedLinkModel'
...
components:
  schemas:
...
    SchemaModel:
      type: object
      required:
        - name
        - attributes
      properties:
        name:
          type: string
        attributes:
          type: array
          items:
            $ref: '#/definitions/AttributeDefinitionModel'
...
    TransactionDefinitionModel:
      type: object
      required:
        - name
        - key
        - processDefinitionKey
        - entitySchema
        - defaultStatus
        - taskFormMappings
        - category
      properties:
        id:
          type: string
          format: uuid
        key:
          type: string
        name:
          type: string
        category:
          type: string
        processDefinitionKey:
          type: string
        entitySchema:
          type: string
        defaultStatus:
          type: string
        taskFormMappings:
          type: array
          items:
            $ref: '#/definitions/TaskFormMappingModel'
        namedFormMappings:
          type: array
          items:
            $ref: '#/definitions/NamedFormMappingModel'
...
    FormConfigDefinitionModel:
      type: object
      required:
        - name
        - schema
        - version
        - status
        - formConfigJson
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        schema:
          type: string
        formConfigJson:
          type: object
          additionalProperties: true
        rendererOptions:
          type: object
          additionalProperties: true
        description:
          type: string
        version:
          type: string
        status:
          type: string
          enum:
            - draft
            - published
        createdBy:
          type: string
        lastUpdatedBy:
          type: string
        translationRequired:
          type: boolean
        createdTimeStamp:
          type: string
          format: date-time
        lastUpdatedTimeStamp:
          type: string
          format: date-time
    FormConfigQueryParameters:
      type: object
      properties:
        id:
          type: string
          format: uuid
...
    TransactionLinkTypeModel:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        fromDescription:
          type: string
        toDescription:
          type: string
    AllowedLinkModel:
      type: object
      properties:
        id:
          type: string
          format: uuid
        transactionDefinitionKey:
          type: string
        transactionLinkType:
          $ref: '#/components/schemas/TransactionLinkTypeModel'
    AllowedLinkCreationRequest:
      type: object
      properties:
        transactionDefinitionKey:
          type: string
        transactionLinkTypeId:
          type: string
          format: uuid
...